Moved player to lobby
