+++ Assets/Scenes/Arena/ArenaScene.unity
@@ -88,7 +88,7 @@ LightmapSettings:
     m_PVRFilteringAtrousPositionSigmaIndirect: 2
     m_PVRFilteringAtrousPositionSigmaAO: 1
     m_ShowResolutionOverlay: 1
-  m_LightingDataAsset: {fileID: 112000000, guid: f26de9d6f766f224c9627a9c8ffda017,
+  m_LightingDataAsset: {fileID: 112000004, guid: f26de9d6f766f224c9627a9c8ffda017,
     type: 2}
   m_UseShadowmask: 1
 --- !u!196 &4
@@ -168,7 +168,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 179325712}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 85962155cbdb67249911e80044e692cd, type: 3}
+  m_Script: {fileID: 11500000, guid: 7b185d7d7a4d2934d9b31cbf264ca9e4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &179325716
@@ -351,7 +351,7 @@ PrefabInstance:
     - target: {fileID: 6355714995786922168, guid: e921085373dbc384a9119a2a4d2989d6,
         type: 3}
       propertyPath: m_LocalPosition.z
-      value: -20
+      value: -95.2
       objectReference: {fileID: 0}
     - target: {fileID: 6355714995786922168, guid: e921085373dbc384a9119a2a4d2989d6,
         type: 3}
@@ -424,6 +424,11 @@ PrefabInstance:
       propertyPath: isGlobal
       value: 1
       objectReference: {fileID: 0}
+    - target: {fileID: 6355714995784231326, guid: e921085373dbc384a9119a2a4d2989d6,
+        type: 3}
+      propertyPath: m_Materials.Array.data[0]
+      value: 
+      objectReference: {fileID: 2100000, guid: 3a0ff0a9266fd0045b6da6f4b1c190d2, type: 2}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: e921085373dbc384a9119a2a4d2989d6, type: 3}
 --- !u!1001 &684060913
@@ -508,10 +513,6 @@ PrefabInstance:
       propertyPath: m_LocalScale.z
       value: 0.2
       objectReference: {fileID: 0}
-    - target: {fileID: 430059796, guid: 2a030bff78f5b684d944779b6139ef0d, type: 3}
-      propertyPath: Animator
-      value: 
-      objectReference: {fileID: 1566769637}
     - target: {fileID: 6090655161758902092, guid: 2a030bff78f5b684d944779b6139ef0d,
         type: 3}
       propertyPath: Animator
@@ -532,6 +533,10 @@ PrefabInstance:
       propertyPath: Rotation.z
       value: 77.339
       objectReference: {fileID: 0}
+    - target: {fileID: 430059796, guid: 2a030bff78f5b684d944779b6139ef0d, type: 3}
+      propertyPath: Animator
+      value: 
+      objectReference: {fileID: 1566769637}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 2a030bff78f5b684d944779b6139ef0d, type: 3}
 --- !u!1001 &821864974
@@ -665,6 +670,21 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 682019370}
   m_PrefabAsset: {fileID: 0}
+--- !u!114 &857522590
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 857522580}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e6aa85e323a0b934eaecb6e0eb764703, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  HasHelmet: 1
+  HasSkirt: 1
+  HasChest: 1
 --- !u!1 &875419017
 GameObject:
   m_ObjectHideFlags: 0
@@ -841,36 +861,12 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
---- !u!1 &1107737528
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1107737529}
-  m_Layer: 0
-  m_Name: Audience
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &1107737529
+--- !u!4 &1107737529 stripped
 Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
+  m_CorrespondingSourceObject: {fileID: 1107737529, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
+    type: 3}
+  m_PrefabInstance: {fileID: 3785299402449313554}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1107737528}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 3785299402449313555}
-  m_RootOrder: 4
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1391897302
 GameObject:
   m_ObjectHideFlags: 0
@@ -1256,44 +1252,8 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 3785299401517316088, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-        type: 3}
-      propertyPath: m_LocalScale.x
-      value: 2.6
-      objectReference: {fileID: 0}
-    - target: {fileID: 3803165561468267057, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -0.08
-      objectReference: {fileID: 0}
-    - target: {fileID: 3803165561468267057, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-        type: 3}
-      propertyPath: m_LocalScale.x
-      value: 9.98
-      objectReference: {fileID: 0}
-    - target: {fileID: 3803165561468267057, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-        type: 3}
-      propertyPath: m_LocalScale.y
-      value: 13.16
-      objectReference: {fileID: 0}
-    - target: {fileID: 3803165561468267057, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-        type: 3}
-      propertyPath: m_LocalScale.z
-      value: 34.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3803165561468267057, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 1.21
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb, type: 3}
---- !u!4 &3785299402449313555 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 3785299401517316088, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
-    type: 3}
-  m_PrefabInstance: {fileID: 3785299402449313554}
-  m_PrefabAsset: {fileID: 0}
 --- !u!95 &5823257552893570132 stripped
 Animator:
   m_CorrespondingSourceObject: {fileID: 5823257552893570132, guid: 87f58c77e0d89374f9ae8c9ecbc1aadb,
