+++ Assets/Scenes/Arena/ArenaScene/LightingData.asset.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: f26de9d6f766f224c9627a9c8ffda017
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 112000000
+  mainObjectFileID: 6475696996188705980
   userData: 
   assetBundleName: 
   assetBundleVariant:
