+++ Assets/Scenes/Arena/ArenaScene/Lightmap-0_comp_dir.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6934eed726ba1137e2f840335538466dce8ab4022867c9e206214990a3c0f5ed
+size 435290
