+++ Assets/Scenes/Arena/ArenaScene/Lightmap-0_comp_light.exr
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4ca9994d3a7b17d3363a1c2a4eb3e13e39531b3db796f8221e08ac098ad537f3
+size 1490122
