+++ Assets/Scenes/Arena/ArenaScene/Lightmap-0_comp_shadowmask.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:35e14a801acbb5f673672cbfbe96350c55d4f39a4013e8f9c5ebbb1069692af3
+size 123921
