--- Assets/Scripts/Game Managers/AudienceManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 098335a3a6d9d9442b4339f094c8809d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
