--- Assets/Scripts/Game Managers/BattleManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85962155cbdb67249911e80044e692cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
