+++ Assets/Scripts/GameManagers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d1cdaf955aaa5a4794b393febf83587
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
