+++ Assets/Scripts/GameManagers/BattleManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b185d7d7a4d2934d9b31cbf264ca9e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
