+++ Assets/Scripts/GameManagers/GateManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 72080152baecdb644bec382c0b899881
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
